uint16

+++ pipeline/backend/kubernetes/service.go
@@ -20,7 +20,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
-func Service(namespace, name, podName string, ports []int32) (*v1.Service, error) {
+func Service(namespace, name, podName string, ports []uint16) (*v1.Service, error) {
 	var svcPorts []v1.ServicePort
 	for _, port := range ports {
 		svcPorts = append(svcPorts, v1.ServicePort{
